Edit SQL statement and Fix Dialog
I added the date_due column in the overdue and loaned books query.

Then I changed the combobox on the add book dialog to a JList like we discussed and added the add book button.
I also noticed you called the "Add Author" button addUserDialogButton and it is in the global class scope instead of in the inner class.
Maybe we should change the name and put it in the inner class.